zipArchive/requirements.txt
amqp==5.1.1
ansicon==1.89.0
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
black==22.8.0
blessed==1.19.1
bpython==0.23
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
curtsies==0.4.0
cwcwidth==0.1.7
Deprecated==1.2.13
Django==4.1.1
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-countries==7.4.2
django-filter==22.1
django-reset-migrations==0.4.0
djangorestframework==3.14.0
dnspython==2.2.1
eventlet==0.33.2
greenlet==1.1.3
idna==3.4
jinxed==1.2.0
kombu==5.2.4
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.32
psycopg2==2.9.5
Pygments==2.13.0
pyparsing==3.0.9
pytz==2022.6
pyxdg==0.28
redis==4.3.4
redis-cli==0.1.2
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.4
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
